refactor: update shift model and API to use new data structure with Instant timestamps

+++ app/src/main/java/com/avoqado/pos/core/data/network/models/transactions/summary/Summary.kt
@@ -10,7 +10,7 @@ data class Summary(
     @SerializedName("ratingsCount")
     val ratingsCount: Int?,
     @SerializedName("totalSales")
-    val totalSales: Int?,
+    val totalSales: Double?,
     @SerializedName("totalTips")
-    val totalTips: Int?,
+    val totalTips: Double?,
 )
