+++ app/src/main/java/com/avoqado/pos/core/domain/models/ShiftSummary.kt
@@ -5,6 +5,6 @@ data class ShiftSummary(
     val averageTipPercentage: Double,
     val ordersCount: Int,
     val ratingsCount: Int,
-    val totalSales: Int,
-    val totalTips: Int,
+    val totalSales: Double,
+    val totalTips: Double,
 )
